more refactoring, UsersList, removed unused files
