--- templates/expenses/create.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang=“en”>
-<head>
-  <meta charset=“UTF-8">
-  <title>Expense Created</title>
-</head>
-<body>
-  <h1>Expense Created</h1>
-  <p>Category: {{ expense.category }}</p>
-  <p>Amount: {{ expense.amount }}</p>
-  <p>Date: {{ expense.date }}</p>
-   <p><a href="{% url 'expenses:create_expense' %}">Submit</a></p>
-  <p><a href="{% url 'expenses:new_expense' %}">Create another</a></p>
-</body>
-</html>
